+++ src/pages/MainPage.tsx
@@ -1,10 +1,9 @@
 // 메뉴 전체 보여줄 메인페이지
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import styled from "styled-components";
 import theme from "../styles/theme";
-import Filter from "../components/MainPage/Filter";
 import Menu from "../components/MainPage/Menu";
-import { burger_image } from "../assets/image/asset";
+import { getCategorydMenu } from "../util/api";
 
 export interface iItem {
   menuId: number;
@@ -16,80 +15,60 @@ export interface iItem {
 
 export default function MainPage() {
   const [quantity, setQuantity] = useState<number>(0);
-  const data: iItem[] = [
-    {
-      menuId: 1,
-      menuName: "트리플 치즈 버거",
-      image: burger_image,
-      priceOnly: 5700,
-      allergy: ["pig", "cow", "chicken", "lettuce"],
-    },
-    {
-      menuId: 2,
-      menuName: "맥스파이시 버거",
-      image: burger_image,
-      priceOnly: 7500,
-      allergy: ["pig", "chicken", "lettuce"],
-    },
-    {
-      menuId: 3,
-      menuName: "페퍼로니 치즈 버거",
-      image: burger_image,
-      priceOnly: 5700,
-      allergy: ["cow", "chicken", "lettuce"],
-    },
-    {
-      menuId: 4,
-      menuName: "1955 버거",
-      image: burger_image,
-      priceOnly: 8500,
-      allergy: ["pig", "cow", "chicken", "lettuce"],
-    },
-    {
-      menuId: 5,
-      menuName: "트리플 치즈 버거",
-      image: burger_image,
-      priceOnly: 5700,
-      allergy: ["lettuce"],
-    },
-    {
-      menuId: 6,
-      menuName: "페퍼로니 치즈 버거",
-      image: burger_image,
-      priceOnly: 5700,
-      allergy: ["pig", "tomato", "chicken", "lettuce"],
-    },
-    {
-      menuId: 7,
-      menuName: "페퍼로니 치즈 버거",
-      image: burger_image,
-      priceOnly: 5700,
-      allergy: ["tomato", "cow", "chicken", "lettuce"],
-    },
-    {
-      menuId: 7,
-      menuName: "페퍼로니 치즈 버거",
-      image: burger_image,
-      priceOnly: 5700,
-      allergy: ["pig", "cow", "chicken", "lettuce"],
-    },
-  ];
+  const [menus, setMenus] = useState<iItem[]>([]);
+  const [currentMenu, setCurrentMenu] = useState({
+    recommendation: false,
+    burger: true,
+    snacks: false,
+    drink: false,
+  });
 
-  const [menus, setMenus] = useState<iItem[]>(data);
+  useEffect(() => {
+    getMenu("burger");
+  }, []);
+
+  const getMenu = (categoryName: string) => {
+    getCategorydMenu(categoryName).then((result: { data: React.SetStateAction<iItem[]> }) => setMenus(result.data));
+  };
+
+  const clickRecommendationButton = () => {
+    setCurrentMenu({ ...currentMenu, recommendation: true, burger: false, snacks: false, drink: false });
+    getMenu("recommend");
+  };
+  const clickBurgerButton = () => {
+    setCurrentMenu({ ...currentMenu, recommendation: false, burger: true, snacks: false, drink: false });
+    getMenu("burger");
+  };
+  const clickSnaksButton = () => {
+    setCurrentMenu({ ...currentMenu, recommendation: false, burger: false, snacks: true, drink: false });
+    getMenu("snack");
+  };
+  const clickDrinkButton = () => {
+    setCurrentMenu({ ...currentMenu, recommendation: false, burger: false, snacks: false, drink: true });
+    getMenu("drink");
+  };
+
   return (
     <>
       <MainBackground>
-        <Filter />
+        <CategoryButtonWrapper>
+          <RecommendationButton onClick={clickRecommendationButton} isClicked={currentMenu.recommendation}>
+            추천 메뉴
+          </RecommendationButton>
+          <BurgerButton onClick={clickBurgerButton} isClicked={currentMenu.burger}>
+            버거&세트
+          </BurgerButton>
+          <SnackButton onClick={clickSnaksButton} isClicked={currentMenu.snacks}>
+            스낵&사이드
+          </SnackButton>
+          <DrinkButton onClick={clickDrinkButton} isClicked={currentMenu.drink}>
+            음료
+          </DrinkButton>
+        </CategoryButtonWrapper>
         <MenuWrapper>
           {menus.map((item: iItem) => {
             return (
-              <Menu
-                menuName={item.menuName}
-                image={item.image}
-                priceOnly={item.priceOnly}
-                allergy={item.allergy}
-                key={item.menuId}
-              />
+              <Menu menuName={item.menuName} priceOnly={item.priceOnly} allergy={item.allergy} key={item.menuId} />
             );
           })}
         </MenuWrapper>
@@ -103,6 +82,52 @@ const MainBackground = styled.div`
   width: 37.5rem;
   height: auto;
   background-color: ${theme.colors.bg};
+`;
+const CategoryButtonWrapper = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  padding: 1.6rem 1rem 1.6rem 2rem;
+`;
+
+const RecommendationButton = styled.button<{ isClicked: boolean }>`
+  width: 7.4rem;
+  height: 4rem;
+  background-color: ${({ isClicked }) => (isClicked ? `${theme.colors.red}` : "#d9d9d9")};
+  ${theme.fonts.body4};
+  color: ${theme.colors.white};
+  padding: 1rem;
+  border-radius: 2rem;
+`;
+
+const BurgerButton = styled.button<{ isClicked: boolean }>`
+  width: 8.7rem;
+  height: 4rem;
+  background-color: ${({ isClicked }) => (isClicked ? `${theme.colors.red}` : "#d9d9d9")};
+  ${theme.fonts.body4};
+  color: ${theme.colors.white};
+  padding: 1rem;
+  border-radius: 2rem;
+`;
+
+const SnackButton = styled.button<{ isClicked: boolean }>`
+  width: 9.9rem;
+  height: 4rem;
+  background-color: ${({ isClicked }) => (isClicked ? `${theme.colors.red}` : "#d9d9d9")};
+  ${theme.fonts.body4};
+  color: ${theme.colors.white};
+  padding: 1rem;
+  border-radius: 2rem;
+`;
+
+const DrinkButton = styled.button<{ isClicked: boolean }>`
+  width: 4.6rem;
+  height: 4rem;
+  background-color: ${({ isClicked }) => (isClicked ? `${theme.colors.red}` : "#d9d9d9")};
+  ${theme.fonts.body4};
+  color: ${theme.colors.white};
+  padding: 1rem;
+  border-radius: 2rem;
 `;
 
 const MenuWrapper = styled.div`
